\name{agescales}
\alias{agescales}
\docType{data}
\title{
Age scales.
}
\description{
Cenozoic paleomagnetic age scales.
}
\usage{data(agescales)}
\format{
  A data frame with 182 observations on the following 4 variables.
  \describe{
    \item{\code{chron}}{Factor. Magnetostratigraphic data.}
    \item{\code{B85}}{Numeric. Age of the datum on Berggren et al. 1985 scale.}
    \item{\code{B95}}{Numeric. Age of the datum on Berggren et al. 1995 scale.}
    \item{\code{G04}}{Numeric. Age of the datum on Gradstein et al. 2004 scale.}
  }
}

\source{
Berggren, W. A:, D. V. Kent, J. J. Flynn and J. A. van Couvering. 1985. Cenozoic Geochronology. GSA Bulletin, 96: 1407-1418.
Berggren, W. A., D. V. Kent, C. C. Swisher III, and M.-P. Aubry. 1995. A revised Cenozoic geochronology and chronostratigra- phy. In W. A. Berggren et al., eds. Geochronology time scales and global stratigraphic correlation. SEPM Special Publication 54:129–212.
Gradstein, F. M., J. G. Ogg, A. G. Smith, F. P. Agterberg, W. Bleeker, R. A. Cooper, V. Davydov, P. Gibbard, L. A. Hinnov, M. R. House, L. Lourens, H.-P. Luterbacher, J. McArthur, M. J. Melchin, L. J. Robb, P. M. Sadler, J. Shergold, M. Villeneuve, B. R. Wardlaw, J. Ali, H. Brinkhuis, F. J. Hilgen, J. Hooker, R. J. Howarth, A. H. Knoll, J. Laskar, S. Monechi, J. Powell, K. A. Plumb, I. Raffi, U. Röhl, A. Sanfilippo, B. Schmitz, N. J. Shackleton, G. A. Shields, H. Strauss, J. Van Dam, J. Veizer, T. van Kolfschoten and D. Watson. 2004. A Geologic Time Scale 2004. Cambridge University Press, 589pp.
}

\examples{
data(agescales)
}
\keyword{datasets}
